+++ packages/backend/src/lib/chat-message.ts
