+++ packages/backend/src/lib/command.ts
