+++ packages/backend/src/services/chains/conversational-chain/conversational-chain.service.ts
@@ -4,10 +4,10 @@ import { BotType, Conversation, KeyNotSetException } from '@my-monorepo/shared';
 import { CallbackManager } from 'langchain/callbacks';
 import createLlm from '../../llm/create-llm';
 import { BufferMemory } from 'langchain/memory';
-import ConversationalChain from '../../../models/chains/conversational.chain';
+import ConversationalChain from 'src/lib/chains/conversational.chain';
 import { ChatHistoryBuilderService } from '../../chat-history-builder/chat-history-builder.service';
-import { ConversationalChainBuilder } from '../../../models/chain-builder';
-import { createAgent } from 'src/models/chains/agent';
+import { ConversationalChainBuilder } from 'src/lib/chain-builder';
+import { createAgent } from 'src/lib/chains/agent';
 
 @Injectable()
 export class ConversationalChainService extends ConversationalChainBuilder {
